orgnised using router and added datalayer and controller
